src/ingestion/requirements.txt
requests==2.31.0
google-cloud-pubsub==2.19.0
pytest==7.4.0
mock==5.1.0